Refactor VxlanTopology
- move static computation functionality into separate class
